feature: 로직 및 컴포넌트 분리 (#11)
