+++ src/hooks/play-list/index.ts
@@ -1,4 +1,5 @@
-export * from './useDragAndDrop';
+export * from './useItemDragAndDrop';
 export * from './useHandleCreateHashtag';
 export * from './useHandleCreatePlaylist';
 export * from './useHandleCreatePlaylistVideo';
+export * from './useBottomSheetDragAndDrop';
