+++ src/pages/PlayListEditPage/PlayListEditPage.tsx
@@ -8,8 +8,12 @@ import type { PlayListEditFormValues } from '@/types';
 import { useHandleCreatePlaylist } from '@/hooks/play-list/useHandleCreatePlaylist';
 import { useAuthStateChange } from '@/hooks';
 import { useFetchCategoryByCategoryNameEn } from '@/hooks/useCategory';
+import { playListAtom, playlistErrorAtom } from '@/atoms';
+import { useAtom, useSetAtom } from 'jotai';
 
 const PlaylistEditPage = () => {
+  const [playlists] = useAtom(playListAtom);
+  const setPlaylistError = useSetAtom(playlistErrorAtom);
   const { user } = useAuthStateChange();
   console.log(user);
   const methods = useForm<PlayListEditFormValues>({
@@ -26,6 +30,12 @@ const PlaylistEditPage = () => {
     useFetchCategoryByCategoryNameEn(watchedCategory);
 
   const handleOnSubmit = async (data: PlayListEditFormValues) => {
+    console.log('호출!');
+    if (playlists.length === 0) {
+      setPlaylistError('영상을 추가해주세요.');
+      return;
+    }
+
     try {
       if (categoryData && categoryData.length > 0) {
         data.category = categoryData[0].category_id;
